refactor(dashboard): extract the nav bar's body into a hook
