+++ src/module/dashboard/ui/nav-bar/nav-bar.tsx
@@ -1,31 +1,23 @@
 'use client';
 
-import { useId, useState } from 'react';
-
 import { BrandLabel } from '@/ui/brand-label/brand-label';
 import { Spinner } from '@/ui/decorative/spinner/spinner';
 import { HamburgerButton } from '@/ui/hamburger-button/hamburger-button';
 import { NavBar } from '@/ui/nav-bar/nav-bar';
-import { useSessionUser } from '@/user/presentation/hook/use-session-user';
 import { UserBadge } from '@/user/presentation/ui/badge/badge';
 
 import { NavBarNav } from './nav/nav';
+import { useDashboardNavBar } from './use-nav-bar';
 
 export function DashboardNavBar() {
-  const { data, isPending } = useSessionUser();
-
-  const navId = useId();
-  const [isActive, setIsActive] = useState(false);
-
-  const handleHamburgerButtonClick = () => {
-    setIsActive((currentState) => !currentState);
-  };
-
-  const handleNavClick = () => {
-    if (isActive) {
-      setIsActive(false);
-    }
-  };
+  const {
+    user,
+    isPending,
+    navId,
+    isActive,
+    handleHamburgerButtonClick,
+    handleNavClick,
+  } = useDashboardNavBar();
 
   return (
     <NavBar>
@@ -33,7 +25,7 @@ export function DashboardNavBar() {
       {isPending && (
         <Spinner className="fill-accent-400 stroke-accent-400 z-10 h-full" />
       )}
-      {data && <UserBadge className="z-10" user={data} />}
+      {user && <UserBadge className="z-10" user={user} />}
       <HamburgerButton
         aria-controls={navId}
         aria-expanded={isActive}
